extracted dialogs
